+++ test/promise-api.test.js
@@ -174,7 +174,7 @@ describe('Promise API', function() {
     });
   });
 
-  it('should run a hamster horde', function(done) {
+  it('should start a hamster horde', function(done) {
     p = pipsqueak([
       { factory: factory, interval: '100ms', },
       { factory: factory, interval: '50ms', },
@@ -183,6 +183,93 @@ describe('Promise API', function() {
       assert.equal(executions, 8);
       done();
     }, 250);
+  });
+
+  it('should poke a hamster horde', function(done) {
+    p = pipsqueak([
+      { factory: factory, interval: '100ms', },
+      { factory: factory, interval: '50ms', },
+    ]).poke();
+    setTimeout(function() {
+      assert.equal(executions, 2);
+      done();
+    }, 250);
+  });
+
+  it('should poke a subset of a hamster horde', function(done) {
+    p = pipsqueak([
+      { name: 'rod', factory: factory, interval: '50ms', },
+      { name: 'jane', factory: factory, interval: '50ms', },
+      { name: 'freddy', factory: factory, interval: '50ms', },
+    ]).poke(['rod', 'jane',]);
+    setTimeout(function() {
+      assert.equal(executions, 2);
+      done();
+    }, 250);
+  });
+
+  it('should poke a single hamster in a hamster horde', function(done) {
+    p = pipsqueak([
+      { name: 'rod', factory: factory, interval: '50ms', },
+      { name: 'jane', factory: factory, interval: '50ms', },
+      { name: 'freddy', factory: factory, interval: '50ms', },
+    ]).poke('rod');
+    setTimeout(function() {
+      assert.equal(executions, 1);
+      done();
+    }, 250);
+  });
+
+  it('should resume existing schedule after being poked', function(done) {
+    p = pipsqueak([
+      { name: 'rod', factory: factory, interval: '100ms', },
+      { name: 'jane', factory: factory, interval: '100ms', },
+      { name: 'freddy', factory: factory, interval: '100ms', },
+    ]).start();
+    setTimeout(function() {
+      assert.equal(executions, 3);
+      p.poke('rod');
+      setTimeout(function() {
+        assert.equal(executions, 4);
+        setTimeout(function() {
+          assert.equal(executions, 6);
+          done();
+        }, 50);
+      }, 25);
+    }, 50);
+  });
+
+  it('should not poke disabled tasks', function(done) {
+    p = pipsqueak([
+      { factory: factory, interval: '50ms', disabled: true, },
+    ]).poke();
+    setTimeout(function() {
+      assert.equal(executions, 0);
+      done();
+    }, 100);
+  });
+
+  it('should not poke stopped tasks', function(done) {
+    p = pipsqueak([
+      { factory: factory, interval: '50ms', },
+    ]);
+    p.stop().then(function() {
+      p.poke();
+      setTimeout(function() {
+        assert.equal(executions, 0);
+        done();
+      }, 100);
+    });
+  });
+
+  it('should not poke running tasks', function(done) {
+    p = pipsqueak([
+      { factory: slow, interval: '50ms', },
+    ]).start().poke();
+    setTimeout(function() {
+      assert.equal(executions, 1);
+      done();
+    }, 100);
   });
 
 });