+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 2.6.0
+### Updated
+- Fixed a bug where if any hamster was running, all hamsters were reported as running
+- Added the ability to poke one or more hamsters
+
 ## 2.5.2
 ### Updated
 - Dependencies